[painter/gl] move texture helper methods into draw.go
