+++ internal/painter/gl/texture.go
@@ -3,6 +3,8 @@ package gl
 import (
 	"fmt"
 	"image"
+	"image/draw"
+	"math"
 
 	"github.com/goki/freetype/truetype"
 
@@ -18,6 +20,17 @@ var noTexture = Texture(cache.NoTexture)
 // Texture represents an uploaded GL texture
 type Texture cache.TextureType
 
+func (p *painter) freeTexture(obj fyne.CanvasObject) {
+	texture, ok := cache.GetTexture(obj)
+	if !ok {
+		return
+	}
+
+	p.ctx.DeleteTexture(Texture(texture))
+	p.logError()
+	cache.DeleteTexture(obj)
+}
+
 func (p *painter) getTexture(object fyne.CanvasObject, creator func(canvasObject fyne.CanvasObject) Texture) (Texture, error) {
 	texture, ok := cache.GetTexture(object)
 
@@ -29,6 +42,48 @@ func (p *painter) getTexture(object fyne.CanvasObject, creator func(canvasObject
 		return noTexture, fmt.Errorf("no texture available")
 	}
 	return Texture(texture), nil
+}
+
+func (p *painter) imgToTexture(img image.Image, textureFilter canvas.ImageScale) Texture {
+	switch i := img.(type) {
+	case *image.Uniform:
+		texture := p.newTexture(textureFilter)
+		r, g, b, a := i.RGBA()
+		r8, g8, b8, a8 := uint8(r>>8), uint8(g>>8), uint8(b>>8), uint8(a>>8)
+		data := []uint8{r8, g8, b8, a8}
+		p.ctx.TexImage2D(
+			texture2D,
+			0,
+			1,
+			1,
+			colorFormatRGBA,
+			unsignedByte,
+			data,
+		)
+		p.logError()
+		return texture
+	case *image.RGBA:
+		if len(i.Pix) == 0 { // image is empty
+			return noTexture
+		}
+
+		texture := p.newTexture(textureFilter)
+		p.ctx.TexImage2D(
+			texture2D,
+			0,
+			i.Rect.Size().X,
+			i.Rect.Size().Y,
+			colorFormatRGBA,
+			unsignedByte,
+			i.Pix,
+		)
+		p.logError()
+		return texture
+	default:
+		rgba := image.NewRGBA(image.Rect(0, 0, img.Bounds().Dx(), img.Bounds().Dy()))
+		draw.Draw(rgba, rgba.Rect, img, image.Point{}, draw.Over)
+		return p.imgToTexture(rgba, textureFilter)
+	}
 }
 
 func (p *painter) newGlCircleTexture(obj fyne.CanvasObject) Texture {
@@ -125,4 +180,32 @@ func (p *painter) newGlTextTexture(obj fyne.CanvasObject) Texture {
 
 	paint.DrawString(img, text.Text, color, face, height, text.TextStyle.TabWidth)
 	return p.imgToTexture(img, canvas.ImageScaleSmooth)
+}
+
+func (p *painter) newTexture(textureFilter canvas.ImageScale) Texture {
+	if int(textureFilter) >= len(textureFilterToGL) {
+		fyne.LogError(fmt.Sprintf("Invalid canvas.ImageScale value (%d), using canvas.ImageScaleSmooth as default value", textureFilter), nil)
+		textureFilter = canvas.ImageScaleSmooth
+	}
+
+	texture := p.ctx.CreateTexture()
+	p.logError()
+	p.ctx.ActiveTexture(texture0)
+	p.ctx.BindTexture(texture2D, texture)
+	p.logError()
+	p.ctx.TexParameteri(texture2D, textureMinFilter, textureFilterToGL[textureFilter])
+	p.ctx.TexParameteri(texture2D, textureMagFilter, textureFilterToGL[textureFilter])
+	p.ctx.TexParameteri(texture2D, textureWrapS, clampToEdge)
+	p.ctx.TexParameteri(texture2D, textureWrapT, clampToEdge)
+	p.logError()
+
+	return texture
+}
+
+func (p *painter) textureScale(v float32) float32 {
+	if p.pixScale == 1.0 {
+		return float32(math.Round(float64(v)))
+	}
+
+	return float32(math.Round(float64(v * p.pixScale)))
 }
